General: Create components for app bar links
