+++ ui/src/app-bar/app-bar.tsx
@@ -1,20 +1,15 @@
 import * as React from 'react';
-import { NavLink } from 'react-router-dom';
 import styles from './app-bar.scss';
 import geoviiteLogo from 'geoviite-design-lib/geoviite-logo.svg';
 //import vaylaLogo from 'vayla-design-lib/logo/vayla-logo.svg';
 import { EnvRestricted } from 'environment/env-restricted';
 import { Environment } from 'environment/environment-info';
 import { useTranslation } from 'react-i18next';
-import { useInfraModelAppSelector, useTrackLayoutAppSelector } from 'store/hooks';
-import { InfraModelTabType } from 'infra-model/infra-model-slice';
-import { InfraModelLink } from 'app-bar/infra-model-link';
-import { getPVDocumentCount } from 'infra-model/infra-model-api';
-import { useLoader } from 'utils/react-utils';
-import { getChangeTimes } from 'common/change-time-api';
+import { InfraModelLink } from 'app-bar/links/infra-model-link';
 import DataProductsMenu from 'app-bar/data-products-menu';
-import { exhaustiveMatchingGuard } from 'utils/type-utils';
 import AppBarMoreMenu from 'app-bar/app-bar-more-menu';
+import { FrontpageLink } from 'app-bar/links/frontpage-link';
+import { AppBarLink } from 'app-bar/links/app-bar-link';
 
 type Link = {
     link: string;
@@ -39,89 +34,17 @@ const links: Link[] = [
 
 export const AppBar: React.FC = () => {
     const { t } = useTranslation();
-    const selectedInfraModelTab = useInfraModelAppSelector((state) => state.infraModelActiveTab);
-    const changeTimes = getChangeTimes();
-    const pvDocumentCounts = useLoader(() => getPVDocumentCount(), [changeTimes.pvDocument]);
-    const exclamationPointVisibility = !!pvDocumentCounts && pvDocumentCounts?.suggested > 0;
-
-    const selectedPublicationId = useTrackLayoutAppSelector(
-        (state) => state.selection.publicationId,
-    );
-
-    const selectedPublicationSearch = useTrackLayoutAppSelector(
-        (state) => state.selection.publicationSearch,
-    );
-
-    function getFrontpageLink(): string {
-        if (selectedPublicationId) {
-            return `/publications/${selectedPublicationId}`;
-        } else if (selectedPublicationSearch) {
-            return '/publications';
-        }
-
-        return `/`;
-    }
-
-    function getInfraModelLink(): string {
-        switch (selectedInfraModelTab) {
-            case InfraModelTabType.PLAN:
-                return '/infra-model/plans';
-            case InfraModelTabType.WAITING:
-                return '/infra-model/waiting-for-approval';
-            case InfraModelTabType.REJECTED:
-                return '/infra-model/rejected';
-            default:
-                return exhaustiveMatchingGuard(selectedInfraModelTab);
-        }
-    }
-
-    function getInfraModelLinkClassName(isActive: boolean): string {
-        return exclamationPointVisibility
-            ? `${styles['app-bar__link']} ${
-                  styles['app-bar__link--infra-model-with-exclamation-point']
-              } ${isActive ? styles['app-bar__link--active'] : ''}`
-            : `${styles['app-bar__link']} 
-             ${isActive ? styles['app-bar__link--active'] : ''}`;
-    }
 
     function getNavLink(link: Link) {
         switch (link.link) {
             case '/':
-                return (
-                    <NavLink
-                        to={getFrontpageLink()}
-                        className={({ isActive }) =>
-                            `${styles['app-bar__link']} ${
-                                isActive ? styles['app-bar__link--active'] : ''
-                            }`
-                        }
-                        end>
-                        {t(link.name)}
-                    </NavLink>
-                );
+                return <FrontpageLink linkTranslation={t(link.name)} />;
+
             case '/infra-model':
-                return (
-                    <NavLink
-                        to={getInfraModelLink()}
-                        className={({ isActive }) => getInfraModelLinkClassName(isActive)}
-                        end>
-                        <InfraModelLink exclamationPointVisibility={exclamationPointVisibility} />
-                    </NavLink>
-                );
+                return <InfraModelLink />;
 
             default:
-                return (
-                    <NavLink
-                        to={link.link}
-                        className={({ isActive }) =>
-                            `${styles['app-bar__link']} ${
-                                isActive ? styles['app-bar__link--active'] : ''
-                            }`
-                        }
-                        end>
-                        {t(link.name)}
-                    </NavLink>
-                );
+                return <AppBarLink linkAddress={link.link} linkName={t(link.name)} />;
         }
     }
